refactor: move AppStore pane methods to +Panes extension
